producer/requirements.txt
kafka-python==2.0.2
google-api-python-client
dotenv
pandas